requirements.txt
aiofiles==0.8.0
alembic==1.8.1
anyio==3.6.1
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
